Change the input processing to produce floats.  Get rid of the MUSE_SHORT_INPUT_MULT constant (we still assume that phase correct input files have the range 0-1023, but that might not stay that way, and we no longer need the scaling internally to keep things in shorts without too much loss of precision).

+++ musecpp/ControlSignalDecoder.h
@@ -16,7 +16,7 @@ enum MotionInformation {
 };
 
 struct ControlSignalDecoder {
-    explicit ControlSignalDecoder(Logger &log, uint16_t const *data, std::pair<float, float> const &eq);
+    explicit ControlSignalDecoder(Logger &log, float const *data, std::pair<float, float> const &eq);
     void log_control_data() const;
     std::optional<int> field_subsampling_phase_Y;
     std::optional<int> horizontal_motion_vector;
